+++ packages/alchemy/src/Cloudflare/Bridge.ts
@@ -1,2 +1,3 @@
 export * from "./Workers/InferEnv.ts";
 export * from "./Workers/Rpc.ts";
+export * from "./Workers/RpcAsync.ts";
