+++ packages/alchemy/test/Cloudflare/Workers/fixtures/http-api/worker.ts
@@ -1,4 +1,4 @@
-import * as Cloudflare from "@/Cloudflare/index.ts";
+import * as Cloudflare from "alchemy/Cloudflare";
 import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
 import * as Path from "effect/Path";
@@ -6,7 +6,7 @@ import * as Etag from "effect/unstable/http/Etag";
 import * as HttpPlatform from "effect/unstable/http/HttpPlatform";
 import * as HttpRouter from "effect/unstable/http/HttpRouter";
 import * as HttpApiBuilder from "effect/unstable/httpapi/HttpApiBuilder";
-import { Task, TaskApi, TaskNotFound } from "./api.ts";
+import { decodeTask, Task, TaskApi, TaskNotFound } from "./api.ts";
 
 const HttpPlatformStub = Layer.succeed(HttpPlatform.HttpPlatform, {
   fileResponse: () => Effect.die("HttpPlatform.fileResponse not supported"),
@@ -20,6 +20,8 @@ const corsLayer = HttpRouter.cors({
   allowedHeaders: ["Content-Type"],
 });
 
+const Bucket = Cloudflare.R2Bucket("Tasks");
+
 export default class HttpApiTestWorker extends Cloudflare.Worker<HttpApiTestWorker>()(
   "HttpApiTestWorker",
   {
@@ -28,25 +30,36 @@ export default class HttpApiTestWorker extends Cloudflare.Worker<HttpApiTestWork
     compatibility: { date: "2024-09-23", flags: ["nodejs_compat"] },
   },
   Effect.gen(function* () {
-    const tasks = new Map<string, Task>();
+    const tasks = yield* Cloudflare.R2Bucket.bind(Bucket);
 
     const tasksGroup = HttpApiBuilder.group(TaskApi, "Tasks", (handlers) =>
       handlers
-        .handle("getTask", ({ params }) => {
-          const task = tasks.get(params.id);
-          if (!task) {
-            return Effect.fail(new TaskNotFound({ id: params.id }));
-          }
-          return Effect.succeed(task);
-        })
-        .handle("createTask", ({ payload }) =>
-          Effect.sync(() => {
+        .handle(
+          "getTask",
+          Effect.fn(function* ({ params }) {
+            const task = yield* tasks.get(params.id).pipe(
+              Effect.flatMap((data) =>
+                data
+                  ? data.text().pipe(Effect.map((data) => JSON.parse(data)))
+                  : Effect.succeed(undefined),
+              ),
+              Effect.orDie,
+            );
+            if (!task) {
+              return yield* Effect.fail(new TaskNotFound({ id: params.id }));
+            }
+            return decodeTask(task);
+          }),
+        )
+        .handle(
+          "createTask",
+          Effect.fn(function* ({ payload }) {
             const task = new Task({
               id: crypto.randomUUID(),
               title: payload.title,
               completed: false,
             });
-            tasks.set(task.id, task);
+            yield* tasks.put(task.id, JSON.stringify(task)).pipe(Effect.orDie);
             return task;
           }),
         ),
@@ -60,5 +73,5 @@ export default class HttpApiTestWorker extends Cloudflare.Worker<HttpApiTestWork
         HttpRouter.toHttpEffect,
       ),
     };
-  }),
+  }).pipe(Effect.provide(Cloudflare.R2BucketBindingLive)),
 ) {}
