+++ packages/alchemy/test/Cloudflare/Workers/fixtures/rpc-http/group.ts
@@ -1,6 +1,7 @@
 import * as Schema from "effect/Schema";
 import * as Rpc from "effect/unstable/rpc/Rpc";
 import * as RpcGroup from "effect/unstable/rpc/RpcGroup";
+import * as RpcSchema from "effect/unstable/rpc/RpcSchema";
 
 export const PingRpcs = RpcGroup.make(
   Rpc.make("Ping", {
@@ -10,5 +11,9 @@ export const PingRpcs = RpcGroup.make(
   Rpc.make("Slow", {
     payload: { ms: Schema.Number },
     success: Schema.Struct({ slept: Schema.Number }),
+  }),
+  Rpc.make("Count", {
+    payload: { upto: Schema.Number },
+    success: RpcSchema.Stream(Schema.Number, Schema.Never),
   }),
 );
