+++ packages/alchemy/test/Cloudflare/Workers/fixtures/rpc-http/worker.ts
@@ -1,6 +1,8 @@
-import * as Cloudflare from "@/Cloudflare/index.ts";
+import * as Cloudflare from "alchemy/Cloudflare";
+import type { HttpEffect } from "alchemy/Http";
 import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
+import * as Stream from "effect/Stream";
 import * as RpcSerialization from "effect/unstable/rpc/RpcSerialization";
 import * as RpcServer from "effect/unstable/rpc/RpcServer";
 import { PingRpcs } from "./group.ts";
@@ -14,24 +16,36 @@ const handlersLayer = PingRpcs.toLayer({
       n: ++counter,
     })),
   Slow: ({ ms }) => Effect.sleep(`${ms} millis`).pipe(Effect.as({ slept: ms })),
+  Count: ({ upto }) =>
+    Stream.fromReadableStream<number, never>({
+      evaluate: () => {
+        let next = 1;
+        return new ReadableStream<number>({
+          pull(controller) {
+            if (next > upto) {
+              controller.close();
+              return;
+            }
+            controller.enqueue(next++);
+          },
+        });
+      },
+      onError: (cause) => cause as never,
+    }),
 });
 
 export default class RpcHttpTestWorker extends Cloudflare.Worker<RpcHttpTestWorker>()(
   "RpcHttpTestWorker",
   {
     main: import.meta.filename,
-    subdomain: { enabled: true, previewsEnabled: false },
-    compatibility: { date: "2024-09-23", flags: ["nodejs_compat"] },
   },
   Effect.gen(function* () {
     return {
       fetch: RpcServer.toHttpEffect(PingRpcs).pipe(
         Effect.provide(
           Layer.mergeAll(handlersLayer, RpcSerialization.layerNdjson),
-          // ^ ndjson is the canonical RpcServer.toHttpEffect protocol;
-          //   it streams one message per line.
         ),
-      ),
+      ) as unknown as HttpEffect,
     };
   }),
 ) {}
